lambda/agent/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-tools>=0.1.0
boto3>=1.34.0
botocore>=1.34.0
mcp>=0.1.0
